Conflicts:
	cave/com.raytheon.uf.viz.ui.menus/src/com/raytheon/uf/viz/ui/menus/xml/IncludeMenuItem.java

Former-commit-id: 27bc369818c877cb712367b34f087669295bdff1 [formerly 27bc369818c877cb712367b34f087669295bdff1 [formerly 8f71c807be1d604cea51ab522f04f8bbf8c11f83]]
Former-commit-id: 58037cb5d150d390168c7aa78eb37b1fef15bfae
Former-commit-id: 63ea8599d7c991c8353577b3655ee44e906cbf32

+++ cave/com.raytheon.uf.viz.d2d.nsharp/src/com/raytheon/uf/viz/d2d/nsharp/rsc/MdlSndNSharpResourceData.java
@@ -94,7 +94,7 @@ public class MdlSndNSharpResourceData extends D2DNSharpResourceData {
 
     @Override
     protected NcSoundingCube getSoundingCube(NsharpStationInfo stnInfo) {
-        float[][] latLon = { { stnInfo.getLatitude(), stnInfo.getLongitude() } };
+        double[][] latLon = { { stnInfo.getLatitude(), stnInfo.getLongitude() } };
         return NcSoundingQuery.pfcSoundingQueryByLatLon(stnInfo.getReftime()
                 .getTime(), stnInfo.getRangestarttime().getTime(), latLon,
                 stnInfo.getSndType(), NcSoundingLayer.DataType.ALLDATA, false,
